Remove metapacconfig and use flatten for the backendconfig struct
